requirements.txt
einops==0.3.0
lmdb==1.2.1
matplotlib==3.2.2
numpy==1.19.1
opencv_python==4.5.4.60
Pillow==9.2.0
scipy==1.5.2
setuptools==52.0.0.post20210125
tensorboardX==2.5.1
torch==1.9.1+cu111
torchvision==0.10.1+cu111
tqdm==4.51.0
